Extract `RyukResourceReaper`
